Added transition between lean states
This added the transition between rotations as well. This is really close, it just has some problems when drasticly moving between left and right.
